Move app/content/ → app/util/: pure-data level config (closes #1232)
`app/content/` was a non-canonical subfolder violating Principle 0 of the ECS
folder layout (`.squad/decisions.md` § DoD source-text grounding). Only
`components/ entities/ systems/ tags/ util/` are permitted directly under
`app/`. Companion to closed #1223 (`app/rendering/` drift collapse).

`level_content_config.{h,cpp}` is pure constexpr config + extern data tables
(LEVELS, LEVEL_KEYS, DIFFICULTY_NAMES, DIFFICULTY_KEYS) with 4 `constexpr`
validity helpers. No I/O, no registry, no entity coupling. Fits canonical
`util/` role (pure functions, pure data).

Migration:
- `git mv app/content/level_content_config.{h,cpp} app/util/`
- 9 include-path updates across app/ and tests/
- CMakeLists: drop redundant `CONTENT_SOURCES` glob (UTIL_SOURCES already
  globs `app/util/*.cpp` and now picks up the moved TU)
- empty `app/content/` directory removed

Acceptance:
- `app/` now contains only canonical subfolders + top-level glue ✅
- Unity + non-unity builds green with zero warnings (`-Wall -Wextra -Werror`) ✅
- All 902 Catch2 test cases (2957 assertions) pass ✅
