+++ app/ui/screen_controllers/level_select_screen_controller.cpp
@@ -2,7 +2,7 @@
 // Dynamic card/difficulty rendering to be ported to rguilayout in future work.
 
 #include "../../components/game_state.h"
-#include "../../content/level_content_config.h"
+#include "../../util/level_content_config.h"
 #include "../../systems/input.h"
 #include "../../constants.h"
 #include "screen_controller_base.h"
