+++ tests/test_high_score_persistence.cpp
@@ -6,7 +6,7 @@
 #include <string_view>
 
 #include "components/high_score.h"
-#include "content/level_content_config.h"
+#include "util/level_content_config.h"
 #include "systems/high_score_system.h"
 
 namespace {
