+++ tests/test_test_player_system.cpp
@@ -1,7 +1,7 @@
 #include <catch2/catch_test_macros.hpp>
 #include "test_helpers.h"
 #include "systems/test_player.h"
-#include "content/level_content_config.h"
+#include "util/level_content_config.h"
 #include "systems/test_player_session.h"
 #include "systems/session_logger_system.h"
 #include <cstdio>
